Breaking up the user-profile-management index.ts file.
