+++ clients/ng-lens-track/src/app/user-profile-management/profile-selector/profile-selector.component.ts
@@ -80,6 +80,10 @@ export class ProfileSelectorComponent implements OnInit, OnDestroy {
     this.editMode = false;
   }
 
+  delete() {
+
+  }
+
   ngOnInit() {
     this.userProfiles$ = this._store.pipe(select(selectUserProfiles));
     const currentUserProfileSubscription = this._store
