+++ clients/ng-lens-track/src/app/user-profile-management/reducers/models.ts
@@ -1,0 +1,11 @@
+import { UserProfile } from '../models/user-profile';
+
+export interface UserProfileManagementState {
+  userProfiles: UserProfile[];
+  currentUserProfileId: number;
+}
+
+export const initialState: UserProfileManagementState = {
+  userProfiles: [],
+  currentUserProfileId: null,
+};
