+++ clients/ng-lens-track/src/app/user-profile-management/reducers/selectors.ts
@@ -1,0 +1,16 @@
+import { createFeatureSelector, createSelector } from '@ngrx/store';
+import { UserProfileManagementState } from './models';
+
+export const selectUserProfileManagementState = createFeatureSelector<
+  UserProfileManagementState
+>('userProfileManagement');
+
+export const selectUserProfiles = createSelector(
+  selectUserProfileManagementState,
+  state => state.userProfiles
+);
+
+export const selectCurrentUserProfile = createSelector(
+  selectUserProfileManagementState,
+  state => state.currentUserProfileId
+);
